Clip_QnA_Pipeline/requirements.txt
accelerate==0.24.1
aiohttp==3.9.1
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
comm==0.2.0
datasets==2.15.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
lightning-utilities==0.10.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.0
numpy==1.24.1
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==9.3.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.28.1
safetensors==0.4.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tokenizers==0.15.0
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchmetrics==1.2.0
torchvision==0.16.1+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.0
transformers==4.35.2
triton==2.1.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.12
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
